requirements.txt
# Core Dependencies
streamlit==1.31.0
python-dotenv==1.0.0

# Google Generative AI (Gemini)
google-generativeai==0.3.2

# Transformers and ML
transformers==4.36.2
torch==2.1.2
sentencepiece==0.1.99
accelerate==0.25.0

# Data Processing
pandas==2.1.4
numpy==1.26.3

# Visualization 
plotly==5.18.0

# JSON handling
json5==0.9.14

# Optional but recommended
protobuf==4.25.1